requirements.txt
# Core web framework
flask>=2.3.0,<3.0
nest-asyncio>=1.6.0

# LLM and AI libraries
litellm==1.77.4
langchain==1.0.0a4
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0
langchain-anthropic>=0.1.0
openai>=1.0.0,<2.0
tiktoken>=0.5.0,<1.0

# HTTP client
httpx>=0.28.1,<0.29

# Utilities
loguru>=0.7.3,<0.8
diskcache>=5.6.3,<6
json-repair>=0.51.0,<0.52
markdown>=3.6,<4
markdownify>=1.2.0,<1.3
rich>=14.1.0,<15
python-dotenv>=1.0.0
